Move Message coder and decoder to own mod
