+++ p2p/src/net/mock/transport/tcp.rs
@@ -13,31 +13,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{
-    io,
-    net::{IpAddr, SocketAddr},
-};
+use std::net::{IpAddr, SocketAddr};
 
 use async_trait::async_trait;
-use bytes::{Buf, BytesMut};
-use tokio::{
-    io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt},
-    net::{TcpListener, TcpStream},
-};
-use tokio_util::codec::{Decoder, Encoder};
-
-use serialization::{Decode, Encode};
+use tokio::net::{TcpListener, TcpStream};
 
 use crate::{
-    constants::MAX_MESSAGE_SIZE,
     net::{
-        mock::{
-            transport::{MockListener, MockStream, MockTransport},
-            types::Message,
-        },
+        mock::transport::{MockListener, MockStream, MockTransport},
         AsBannableAddress, IsBannableAddress,
     },
-    P2pError, Result,
+    Result,
 };
 
 #[derive(Debug)]
@@ -95,117 +81,6 @@ impl MockListener<TcpMockStream, SocketAddr> for TcpMockListener {
     }
 }
 
-pub type TcpMockStream = TcpStream;
-
-#[async_trait]
-impl MockStream for TcpMockStream {}
-
-struct EncoderDecoder {}
-
-impl Decoder for EncoderDecoder {
-    type Item = Message;
-    type Error = P2pError;
-
-    fn decode(&mut self, src: &mut BytesMut) -> Result<Option<Self::Item>> {
-        if src.len() < 4 {
-            return Ok(None);
-        }
-
-        let mut length_bytes = [0u8; 4];
-        length_bytes.copy_from_slice(&src[..4]);
-        let length = u32::from_le_bytes(length_bytes) as usize;
-
-        if length > MAX_MESSAGE_SIZE {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                format!("Frame of length {length} is too large"),
-            )
-            .into());
-        }
-
-        if src.len() < 4 + length {
-            src.reserve(4 + length - src.len());
-            return Ok(None);
-        }
-
-        let data = src[4..4 + length].to_vec();
-        src.advance(4 + length);
-
-        match Message::decode(&mut &data[..]) {
-            Ok(msg) => Ok(Some(msg)),
-            Err(e) => {
-                Err(std::io::Error::new(std::io::ErrorKind::InvalidData, e.to_string()).into())
-            }
-        }
-    }
-}
-
-impl Encoder<Message> for EncoderDecoder {
-    type Error = P2pError;
-
-    fn encode(&mut self, msg: Message, dst: &mut BytesMut) -> Result<()> {
-        let encoded = msg.encode();
-
-        if encoded.len() > MAX_MESSAGE_SIZE {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                format!("Frame of length {} is too large", encoded.len()),
-            )
-            .into());
-        }
-
-        let len_slice = u32::to_le_bytes(encoded.len() as u32);
-
-        dst.reserve(4 + encoded.len());
-        dst.extend_from_slice(&len_slice);
-        dst.extend_from_slice(&encoded);
-
-        Ok(())
-    }
-}
-
-pub struct EncoderDecoderWithBuf<S> {
-    stream: S,
-    buffer: BytesMut,
-}
-
-impl<S: AsyncWrite + AsyncRead + Unpin> EncoderDecoderWithBuf<S> {
-    pub fn new(stream: S) -> EncoderDecoderWithBuf<S> {
-        EncoderDecoderWithBuf {
-            stream,
-            buffer: BytesMut::new(),
-        }
-    }
-
-    pub async fn send(&mut self, msg: Message) -> Result<()> {
-        let mut buf = bytes::BytesMut::new();
-        EncoderDecoder {}.encode(msg, &mut buf)?;
-        self.stream.write_all(&buf).await?;
-        self.stream.flush().await?;
-        Ok(())
-    }
-
-    /// Read a framed message from socket
-    ///
-    /// First try to decode whatever may be in the stream's buffer and if it's empty
-    /// or the frame hasn't been completely received, wait on the socket until the buffer
-    /// has all data. If the buffer has a full frame that can be decoded, return that without
-    /// calling the socket first.
-    pub async fn recv(&mut self) -> Result<Option<Message>> {
-        loop {
-            match (EncoderDecoder {}.decode(&mut self.buffer)) {
-                Ok(None) => {
-                    if self.stream.read_buf(&mut self.buffer).await? == 0 {
-                        return Err(io::Error::from(io::ErrorKind::UnexpectedEof).into());
-                    }
-                    continue;
-                }
-                frame => return frame,
-            }
-        }
-    }
-}
-
 impl AsBannableAddress for SocketAddr {
     type BannableAddress = IpAddr;
 
@@ -220,12 +95,20 @@ impl IsBannableAddress for SocketAddr {
     }
 }
 
+pub type TcpMockStream = TcpStream;
+
+#[async_trait]
+impl MockStream for TcpMockStream {}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::net::{
         message::{BlockListRequest, Request},
-        mock::types::MockRequestId,
+        mock::{
+            transport::EncoderDecoderWithBuf,
+            types::{Message, MockRequestId},
+        },
     };
 
     #[tokio::test]
